+++ scripts/test-docs.sh
@@ -239,7 +239,7 @@ download_http_artifacts()
       sed 's/^.*\(http[^ ]*\).*$/\1/' |\
       sort |\
       uniq |\
-      grep 'csv\|arff\|bin\|png' |\
+      grep 'csv\|arff\|bin\|png\|jpg\|bz2\|gz' |\
       sed 's/\.$//'`;
   cd $output_dir;
   for a in $artifacts;
@@ -252,6 +252,18 @@ download_http_artifacts()
       then
         echo "Error downloading $a!";
         exit 1;
+      fi
+
+      if [[ $a == *.gz ]];
+      then
+        echo "Unpacking $a...";
+        tar -xzpf *.gz;
+      fi
+
+      if [[ $a == *.bz2 ]];
+      then
+        echo "Unpacking $a...";
+        tar -xjpf *.bz2;
       fi
     fi
   done