+++ src/mlpack/tests/image_test.cpp
@@ -19,8 +19,6 @@ using namespace mlpack;
 using namespace mlpack::data;
 using namespace std;
 
-#ifdef MLPACK_HAS_STB // Compile this only if stb is present.
-
 /**
  * Test if an image with an unsupported extension throws an expected
  * exception.
@@ -30,7 +28,7 @@ TEST_CASE("LoadInvalidExtensionFile", "[ImageLoadTest]")
   arma::Mat<unsigned char> matrix;
   data::ImageInfo info;
 
-  REQUIRE_THROWS_AS(data::Load("invalidExtendion.p4ng", matrix, info,
+  REQUIRE_THROWS_AS(data::Load("invalidExtension.p4ng", matrix, info,
       true),  std::runtime_error);
 }
 
@@ -152,4 +150,89 @@ TEST_CASE("ImageInfoSerialization", "[ImageLoadTest]")
   REQUIRE(info.Quality() == binaryInfo.Quality());
 }
 
-#endif // MLPACK_HAS_STB.
+/**
+ * Test resize the image if this is done correctly.  Try it with a few different
+ * types.
+ */
+TEMPLATE_TEST_CASE("ImagesResizeTest", "[ImageTest]", unsigned char, size_t,
+    float, double)
+{
+  typedef TestType eT;
+
+  arma::Mat<eT> image, images;
+  data::ImageInfo info, resizedInfo, resizedInfo2;
+  std::vector<std::string> files =
+      {"sheep_1.jpg", "sheep_2.jpg", "sheep_3.jpg", "sheep_4.jpg",
+       "sheep_5.jpg", "sheep_6.jpg"};
+  std::vector<std::string> reSheeps =
+      {"re_sheep_1.jpg", "re_sheep_2.jpg", "re_sheep_3.jpg", "re_sheep_4.jpg",
+       "re_sheep_5.jpg", "re_sheep_6.jpg"};
+  std::vector<std::string> smSheeps =
+      {"sm_sheep_1.jpg", "sm_sheep_2.jpg", "sm_sheep_3.jpg", "sm_sheep_4.jpg",
+       "sm_sheep_5.jpg", "sm_sheep_6.jpg"};
+
+  // Load and Resize each one of them individually, because they do not have
+  // the same sizes, and then the resized images, will be used in the next
+  // test.
+  for (size_t i = 0; i < files.size(); i++)
+  {
+    REQUIRE(data::Load(files.at(i), image, info, false) == true);
+    ResizeImages(image, info, 320, 320);
+    REQUIRE(data::Save(reSheeps.at(i), image, info, false) == true);
+  }
+
+  // Since they are all resized, this should passes
+  REQUIRE(data::Load(reSheeps, images, resizedInfo, false) == true);
+
+  REQUIRE(info.Width() == resizedInfo.Width());
+  REQUIRE(info.Height() == resizedInfo.Height());
+
+  REQUIRE(data::Load(reSheeps, images, info, false) == true);
+
+  ResizeImages(images, info, 160, 160);
+
+  REQUIRE(data::Save(smSheeps, images, info, false) == true);
+
+  REQUIRE(data::Load(smSheeps, images, resizedInfo2, false) == true);
+
+  REQUIRE(info.Width() == resizedInfo2.Width());
+  REQUIRE(info.Height() == resizedInfo2.Height());
+
+  // cleanup generated images.
+  for (size_t i = 0; i < reSheeps.size(); ++i)
+  {
+    remove(reSheeps.at(i).c_str());
+    remove(smSheeps.at(i).c_str());
+  }
+}
+
+/**
+ * Test if we resize to the same original dimension we will get the same pixels
+ * and no modification to the image.  Try it with a few different types.
+ */
+TEMPLATE_TEST_CASE("IdenticalResizeTest", "[ImageTest]", unsigned char, size_t,
+    float, double)
+{
+  typedef TestType eT;
+
+  arma::Mat<eT> image;
+  data::ImageInfo info;
+  std::vector<std::string> files =
+      {"sheep_1.jpg", "sheep_2.jpg", "sheep_3.jpg", "sheep_4.jpg",
+       "sheep_5.jpg", "sheep_6.jpg"};
+
+  for (size_t i = 0; i < files.size(); i++)
+  {
+    REQUIRE(data::Load(files.at(i), image, info, false) == true);
+    arma::Mat<eT> originalImage = image;
+    ResizeImages(image, info, info.Width(), info.Height());
+    for (size_t i = 0; i < originalImage.n_rows; ++i)
+    {
+      for (size_t j = 0; j < originalImage.n_cols; ++j)
+      {
+        REQUIRE(originalImage.at(i, j) == image.at(i, j));
+      }
+    }
+  }
+}
+